requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.1
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.8.0
twisted-iocpsupport==1.0.3
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.2
wrapt==1.14.1
zope.interface==5.5.0
